split package and fix sealno issue
